# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:3000/api
  variables: {}
paths:
...
  /jobs:
    post:
      tags:
      - Jobs
      summary: create jobs
      operationId: createjobs
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/createjobsrequest'
              - example:
                  company: ''
                  position: ''
            example:
              company: ''
              position: ''
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Jobs
      summary: Get all jobs
      operationId: Getalljobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    createjobsrequest:
      title: createjobsrequest
      required:
      - company
      - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: ''
        position: ''
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...